--- contrib/net/http/httptrace.go
@@ -1,127 +1,0 @@
-package http
-
-import (
-	"context"
-	"crypto/tls"
-	"net"
-	"net/http/httptrace"
-	"time"
-)
-
-type httpTraceResult struct {
-	DNSLookup        time.Duration
-	TCPConnection    time.Duration
-	TLSHandshake     time.Duration
-	ServerProcessing time.Duration
-	NameLookup       time.Duration
-	Connect          time.Duration
-	Pretransfer      time.Duration
-	StartTransfer    time.Duration
-
-	dnsStart      time.Time
-	dnsDone       time.Time
-	tcpStart      time.Time
-	tcpDone       time.Time
-	tlsStart      time.Time
-	tlsDone       time.Time
-	serverStart   time.Time
-	serverDone    time.Time
-	transferStart time.Time
-
-	isTLS    bool
-	isReused bool
-
-	remoteIP   string
-	remotePort string
-}
-
-func WithClientTrace(ctx context.Context, r *httpTraceResult) context.Context {
-	return httptrace.WithClientTrace(ctx, &httptrace.ClientTrace{
-		DNSStart: func(i httptrace.DNSStartInfo) {
-			r.dnsStart = time.Now()
-		},
-
-		DNSDone: func(i httptrace.DNSDoneInfo) {
-			r.dnsDone = time.Now()
-
-			r.DNSLookup = r.dnsDone.Sub(r.dnsStart)
-			r.NameLookup = r.dnsDone.Sub(r.dnsStart)
-		},
-
-		ConnectStart: func(_, _ string) {
-			r.tcpStart = time.Now()
-
-			if r.dnsStart.IsZero() {
-				r.dnsStart = r.tcpStart
-				r.dnsDone = r.tcpStart
-			}
-		},
-
-		ConnectDone: func(network, addr string, err error) {
-			r.tcpDone = time.Now()
-
-			r.TCPConnection = r.tcpDone.Sub(r.tcpStart)
-			r.Connect = r.tcpDone.Sub(r.dnsStart)
-		},
-
-		TLSHandshakeStart: func() {
-			r.isTLS = true
-			r.tlsStart = time.Now()
-		},
-
-		TLSHandshakeDone: func(_ tls.ConnectionState, _ error) {
-			r.tlsDone = time.Now()
-
-			r.TLSHandshake = r.tlsDone.Sub(r.tlsStart)
-			r.Pretransfer = r.tlsDone.Sub(r.dnsStart)
-		},
-
-		GotConn: func(i httptrace.GotConnInfo) {
-			if i.Reused {
-				r.isReused = true
-			}
-
-			r.remoteIP, r.remotePort, _ = net.SplitHostPort(i.Conn.RemoteAddr().String())
-		},
-
-		WroteRequest: func(info httptrace.WroteRequestInfo) {
-			r.serverStart = time.Now()
-
-			if r.dnsStart.IsZero() && r.tcpStart.IsZero() {
-				now := r.serverStart
-
-				r.dnsStart = now
-				r.dnsDone = now
-				r.tcpStart = now
-				r.tcpDone = now
-			}
-
-			if r.isReused {
-				now := r.serverStart
-
-				r.dnsStart = now
-				r.dnsDone = now
-				r.tcpStart = now
-				r.tcpDone = now
-				r.tlsStart = now
-				r.tlsDone = now
-			}
-
-			if r.isTLS {
-				return
-			}
-
-			r.TLSHandshake = r.tcpDone.Sub(r.tcpDone)
-			r.Pretransfer = r.Connect
-		},
-
-		GotFirstResponseByte: func() {
-			r.serverDone = time.Now()
-
-			r.ServerProcessing = r.serverDone.Sub(r.serverStart)
-			r.StartTransfer = r.serverDone.Sub(r.dnsStart)
-
-			r.transferStart = r.serverDone
-		},
-	})
-}
