+++ contrib/net/http/roundtripper.go
@@ -6,11 +6,16 @@
 package http
 
 import (
+	"context"
+	"crypto/tls"
 	"fmt"
 	"math"
+	"net"
 	"net/http"
+	"net/http/httptrace"
 	"os"
 	"strconv"
+	"time"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
@@ -53,12 +58,11 @@ func (rt *roundTripper) RoundTrip(req *http.Request) (res *http.Response, err er
 	var httpTraceResult httpTraceResult
 	ctx = WithClientTrace(ctx, &httpTraceResult)
 
-	// inject the span context into the http request
 	err = tracer.Inject(span.Context(), tracer.HTTPHeadersCarrier(req.Header))
 	if err != nil {
-		// this should never happen
 		fmt.Fprintf(os.Stderr, "contrib/net/http.Roundtrip: failed to inject http headers: %v\n", err)
 	}
+
 	res, err = rt.base.RoundTrip(req.WithContext(ctx))
 	if err != nil {
 		span.SetTag("http.errors", err.Error())
@@ -110,4 +114,104 @@ func WrapClient(c *http.Client, opts ...RoundTripperOption) *http.Client {
 	}
 	c.Transport = WrapRoundTripper(c.Transport, opts...)
 	return c
+}
+
+type httpTraceResult struct {
+	DNSLookup        time.Duration
+	TCPConnection    time.Duration
+	TLSHandshake     time.Duration
+	ServerProcessing time.Duration
+	NameLookup       time.Duration
+	Connect          time.Duration
+	Pretransfer      time.Duration
+	StartTransfer    time.Duration
+	dnsStart         time.Time
+	dnsDone          time.Time
+	tcpStart         time.Time
+	tcpDone          time.Time
+	tlsStart         time.Time
+	tlsDone          time.Time
+	serverStart      time.Time
+	serverDone       time.Time
+	transferStart    time.Time
+	isTLS            bool
+	isReused         bool
+	remoteIP         string
+	remotePort       string
+}
+
+func WithClientTrace(ctx context.Context, r *httpTraceResult) context.Context {
+	return httptrace.WithClientTrace(ctx, &httptrace.ClientTrace{
+		DNSStart: func(i httptrace.DNSStartInfo) {
+			r.dnsStart = time.Now()
+		},
+		DNSDone: func(i httptrace.DNSDoneInfo) {
+			r.dnsDone = time.Now()
+			r.DNSLookup = r.dnsDone.Sub(r.dnsStart)
+			r.NameLookup = r.dnsDone.Sub(r.dnsStart)
+		},
+		ConnectStart: func(_, _ string) {
+			r.tcpStart = time.Now()
+			if r.dnsStart.IsZero() {
+				r.dnsStart = r.tcpStart
+				r.dnsDone = r.tcpStart
+			}
+		},
+		ConnectDone: func(network, addr string, err error) {
+			r.tcpDone = time.Now()
+			r.TCPConnection = r.tcpDone.Sub(r.tcpStart)
+			r.Connect = r.tcpDone.Sub(r.dnsStart)
+		},
+		TLSHandshakeStart: func() {
+			r.isTLS = true
+			r.tlsStart = time.Now()
+		},
+		TLSHandshakeDone: func(_ tls.ConnectionState, _ error) {
+			r.tlsDone = time.Now()
+			r.TLSHandshake = r.tlsDone.Sub(r.tlsStart)
+			r.Pretransfer = r.tlsDone.Sub(r.dnsStart)
+		},
+		GotConn: func(i httptrace.GotConnInfo) {
+			if i.Reused {
+				r.isReused = true
+			}
+
+			r.remoteIP, r.remotePort, _ = net.SplitHostPort(i.Conn.RemoteAddr().String())
+		},
+
+		WroteRequest: func(info httptrace.WroteRequestInfo) {
+			r.serverStart = time.Now()
+			if r.dnsStart.IsZero() && r.tcpStart.IsZero() {
+				now := r.serverStart
+				r.dnsStart = now
+				r.dnsDone = now
+				r.tcpStart = now
+				r.tcpDone = now
+			}
+
+			if r.isReused {
+				now := r.serverStart
+				r.dnsStart = now
+				r.dnsDone = now
+				r.tcpStart = now
+				r.tcpDone = now
+				r.tlsStart = now
+				r.tlsDone = now
+			}
+
+			if r.isTLS {
+				return
+			}
+
+			r.TLSHandshake = r.tcpDone.Sub(r.tcpDone)
+			r.Pretransfer = r.Connect
+		},
+
+		GotFirstResponseByte: func() {
+			r.serverDone = time.Now()
+			r.ServerProcessing = r.serverDone.Sub(r.serverStart)
+			r.StartTransfer = r.serverDone.Sub(r.dnsStart)
+			r.transferStart = r.serverDone
+		},
+	})
 }
